only search for KEYBIND.dat
fix the 'access incorrect directory' bug by only searching for
KEYBIND.dat since that's all that is needed anyway right now. this
prevents firing the watcher on a subdirectory change, e.g. the creation
or removal of /log/, which was what was causing the issue.
also merge with upstream master